+++ monit/entrypoint.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+env > /opt/monit/scripts/.env
+
+exec "$@"